+++ app/database/models/post.py
@@ -1,4 +1,4 @@
-from sqlalchemy import ForeignKey, Integer, TIMESTAMP, func, String, Boolean
+from sqlalchemy import ForeignKey, BIGINT, TIMESTAMP, func, String, Boolean
 from sqlalchemy.orm import relationship, Mapped, mapped_column
 from datetime import datetime
 
@@ -8,8 +8,8 @@ from .base import Base
 class Post(Base):
     __tablename__ = 'posts'
 
-    id: Mapped[int] = mapped_column(Integer, primary_key=True)
-    user_id: Mapped[int] = mapped_column(Integer, ForeignKey('users.id'), nullable=False, index=True)
+    id: Mapped[int] = mapped_column(BIGINT, primary_key=True)
+    user_id: Mapped[int] = mapped_column(BIGINT, ForeignKey('users.id'), nullable=False, index=True)
     created_at: Mapped[datetime] = mapped_column(TIMESTAMP, server_default=func.now(), nullable=False, index=True)
     updated_at: Mapped[datetime] = mapped_column(TIMESTAMP, server_default=func.now(), onupdate=func.now(), nullable=True)
     text_content: Mapped[str] = mapped_column(String, nullable=True)
